+++ modules/photosynthesis/NEWS.md
@@ -1,5 +1,7 @@
 # PEcAn.photosynthesis 1.7.3.9000
 
+## License change
+* PEcAn.photosynthesis is now distributed under the BSD three-clause license instead of the NCSA Open Source license.
 
 
 # PEcAn.photosynthesis 1.7.3